Use eleventyignore instead of gitignore
We want to generate css from our scss file. Generated css will be inside
src/_includes/css. So we want git to ignore that path but we still want
eleventy to pay attention to it. That's why we edit our gitignore, tell
eleventy not to use gitignore, and add eleventyignore.